change Document.Boxes to be of type IEnumerable<BoxProperties> instead of a List<>
